Reorganize the order of `decode` functions
This is pedantic, but putting the `Optional` versions of these first
made it seem like they were the preferred version, which I'm not sure is
the case. I think at this point we should probably push the `Decoded`
type to people since it's used more commonly in our instance-level
`decode` functions.