Improve the API for root key decoding
I think the original reason for ordering the arguments as `rootKey`,
`object` was to be able to curry the function and partially apply the
root key (as seen in the tests). IMO, this is a weird optimization and
let to a kinda funky API. I think it probably makes more sense to accept
that currying in this way isn't reasonable, and instead order the
arguments so that our `decode` functions have a consistent pattern.